Allow build on Linux without undefined function errors.
If a specific C language standard is given, GNU libc hides all functions that were not defined at the time of the standard's release.
We require some of those functions so we use the less strict gnuxx standards.
The standard restriction in the compiler flags is just for testing anyway.